完成数据排序步骤的优化。主要问题在于数据类型导致缓慢，matrixXd的点云的复制比Xf的慢约1.5倍，而此阶段并不涉及Sum的精度不够问题，用Xf完全没问题。顺便小改了数据排序的代码，用eigen的block复制一格中的一行点云，比手写复制一个一个点云会快一点点。

+++ include/PointsSumPyramid.h
@@ -16,14 +16,14 @@ private:
     int *m_num_elements_each_layer;
     int m_smallest_patch_width;
     PointsSum *m_data;
-    void PreComputeSumOfArray(const Eigen::MatrixXd &cloud_array, PointsSum &points_sum_result);
+    void PreComputeSumOfArray(const Eigen::MatrixXf &cloud_array, PointsSum &points_sum_result);
 
-    int CountJumps(const Eigen::MatrixXd &Z_matrix, float max_diff) const;
+    int CountJumps(const Eigen::MatrixXf &Z_matrix, float max_diff) const;
 public:
     PointsSumPyramid();
     ~PointsSumPyramid();
     void Initialize(const int *num_childs_each_level, int max_index_layers, int smallest_patch_width);
-    void PreComputeSum(const Eigen::MatrixXd &cloud_array);
+    void PreComputeSum(const Eigen::MatrixXf &cloud_array);
     int GetOffsetOfLayerN(int n);
 
     PointsSum &Index(const int ind[], int n);
